pyproject.toml
[tool.poetry]
name = "zeta-discord-bot"
version = "0.1.0"
description = ""
authors = ["Toastx <toastx.sol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
zetamarkets-py = "^0.2.41"
hikari = "^2.0.0.dev122"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
